Feature addition to H2 Database
The objective of this project was to study existing functionality and
add new features to the open source H2 database.
The features added were:
1. The existing implementation of the h2 database allowed to drop only
one column at a time. So in the  added new feature, the code was
modified so that multiple columns could be dropped using a single alter
table query.
2. A new data type 'Enum' was added which provided the functionality to
set a variable value from a set of predefined values.